+++ src/lib/declarationHandler.js
@@ -70,7 +70,9 @@ const CLASSES_OF_TRUTH = [
     'SSHD',
     'Tunnel',
     'TrafficGroup',
-    'Disk'
+    'Disk',
+    'MirrorIp',
+    'RoutingAsPath'
 ];
 
 /**