+++ test/unit/src/lib/configManagerTests.js
@@ -91,7 +91,7 @@ describe('configManager', () => {
             {
                 name: 'route1',
                 partition: 'LOCAL_ONLY',
-                gw: '5.6.7.8',
+                tmInterface: '/Common/myVlan',
                 network: '5.5.5.5',
                 mtu: 1500
             },
@@ -99,7 +99,7 @@ describe('configManager', () => {
                 name: 'outsideDoRoute',
                 partition: 'otherPartition',
                 gw: '3.3.3.5',
-                netowrk: 'default',
+                network: 'default',
                 mtu: 12
             }
         ];
@@ -117,7 +117,7 @@ describe('configManager', () => {
                             network: 'default'
                         },
                         route1: {
-                            gw: '5.6.7.8',
+                            target: 'myVlan',
                             mtu: 1500,
                             name: 'route1',
                             network: '5.5.5.5',
@@ -1393,6 +1393,55 @@ describe('configManager', () => {
             return configManager.get({}, state, doState)
                 .then(() => {
                     assert.deepStrictEqual(state.currentConfig.Common, expectedConfig);
+                });
+        });
+    });
+
+    describe('RoutingAsPath', () => {
+        it('should handle RoutingAsPath with references', () => {
+            const configItems = getConfigItems('RoutingAsPath');
+
+            listResponses['/tm/net/routing/as-path'] = [
+                {
+                    name: 'exampleAsPath',
+                    entriesReference: {
+                        link: 'https://localhost/mgmt/tm/net/routing/as-path/~Common~exampleAsPath/entries?ver=14.1.2.7'
+                    }
+                }
+            ];
+            listResponses['/tm/net/routing/as-path/~Common~exampleAsPath/entries'] = [
+                {
+                    name: '10',
+                    action: 'permit',
+                    regex: '^$'
+                },
+                {
+                    name: '15',
+                    action: 'permit',
+                    regex: '^123'
+                }
+            ];
+
+            const configManager = new ConfigManager(configItems, bigIpMock);
+            return configManager.get({}, state, doState)
+                .then(() => {
+                    assert.deepStrictEqual(state.currentConfig.Common.RoutingAsPath, {
+                        exampleAsPath: {
+                            name: 'exampleAsPath',
+                            entries: [
+                                {
+                                    action: 'permit',
+                                    name: 10,
+                                    regex: '^$'
+                                },
+                                {
+                                    action: 'permit',
+                                    name: 15,
+                                    regex: '^123'
+                                }
+                            ]
+                        }
+                    });
                 });
         });
     });