+++ test/unit/src/lib/networkHandlerTests.js
@@ -63,14 +63,26 @@ describe('networkHandler', () => {
                 return Promise.resolve();
             },
             modify(path, data) {
-                if (!dataSent) {
-                    dataSent = {};
-                }
                 if (!dataSent[path]) {
                     dataSent[path] = [];
                 }
                 dataSent[path].push(data);
                 return Promise.resolve();
+            },
+            transaction(commands) {
+                let promise = Promise.resolve();
+                commands.forEach((command) => {
+                    if (command.method === 'create') {
+                        promise = promise.then(() => this.create(command.path, command.body));
+                    } else if (command.method === 'delete') {
+                        promise = promise.then(() => this.delete(command.path));
+                    } else {
+                        promise = promise.then(() => {
+                            throw new Error(`Unrecognized command method: ${command.method}`);
+                        });
+                    }
+                });
+                return promise;
             }
         };
     });
@@ -758,7 +770,9 @@ describe('networkHandler', () => {
 
             state = {
                 currentConfig: {
-                    Common: {}
+                    Common: {
+                        Route: {}
+                    }
                 }
             };
 
@@ -766,7 +780,6 @@ describe('networkHandler', () => {
             return networkHandler.process()
                 .then(() => {
                     const routeData = dataSent[PATHS.Route];
-                    assert.deepEqual(deletedPaths, []);
                     assert.strictEqual(routeData[0].name, 'route1');
                     assert.strictEqual(routeData[0].gw, '0.0.0.0');
                     assert.strictEqual(routeData[0].network, 'default');
@@ -797,83 +810,6 @@ describe('networkHandler', () => {
                 });
         });
 
-        it('should delete and recreate both Routes if network updated', () => {
-            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
-            return networkHandler.process()
-                .then(() => {
-                    const routeData = dataSent[PATHS.Route];
-                    assert.deepEqual(deletedPaths, [
-                        '/tm/net/route/~Common~theRoute',
-                        '/tm/net/route/~LOCAL_ONLY~localRoute'
-                    ]);
-                    assert.deepStrictEqual(routeData[0], {
-                        gw: '10.11.12.13',
-                        mtu: 1000,
-                        name: 'theRoute',
-                        network: '50.60.70.80/32',
-                        partition: 'Common'
-                    });
-                    assert.deepStrictEqual(routeData[1], {
-                        interface: '/Common/targetVLAN',
-                        mtu: 1120,
-                        name: 'localRoute',
-                        network: '50.60.70.81/32',
-                        partition: 'LOCAL_ONLY'
-                    });
-                });
-        });
-
-        it('should delete only Routes that have their network updated', () => {
-            state.currentConfig.Common.Route.theRoute.network = '50.60.70.80/32';
-            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
-            return networkHandler.process()
-                .then(() => {
-                    const routeData = dataSent[PATHS.Route];
-                    assert.deepEqual(deletedPaths, ['/tm/net/route/~LOCAL_ONLY~localRoute']);
-                    assert.deepStrictEqual(routeData, [
-                        {
-                            gw: '10.11.12.13',
-                            mtu: 1000,
-                            name: 'theRoute',
-                            network: '50.60.70.80/32',
-                            partition: 'Common'
-                        }, {
-                            interface: '/Common/targetVLAN',
-                            mtu: 1120,
-                            name: 'localRoute',
-                            network: '50.60.70.81/32',
-                            partition: 'LOCAL_ONLY'
-                        }
-                    ]);
-                });
-        });
-
-        it('should not delete the existing Route if network not updated', () => {
-            state.currentConfig.Common.Route.theRoute.network = '50.60.70.80/32';
-            state.currentConfig.Common.Route.localRoute.network = '50.60.70.81/32';
-            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
-            return networkHandler.process()
-                .then(() => {
-                    const routeData = dataSent[PATHS.Route];
-                    assert.deepEqual(deletedPaths, []);
-                    assert.deepStrictEqual(routeData, [
-                        {
-                            gw: '10.11.12.13',
-                            mtu: 1000,
-                            name: 'theRoute',
-                            network: '50.60.70.80/32',
-                            partition: 'Common'
-                        }, {
-                            interface: '/Common/targetVLAN',
-                            mtu: 1120,
-                            name: 'localRoute',
-                            network: '50.60.70.81/32',
-                            partition: 'LOCAL_ONLY'
-                        }
-                    ]);
-                });
-        });
-
         it('should not do anything with an empty Route', () => {
             declaration = {
                 Common: {
@@ -897,6 +833,51 @@ describe('networkHandler', () => {
             return networkHandler.process()
                 .then(() => {
                     assert.deepStrictEqual(createdFolder, undefined);
+                });
+        });
+
+        let bigIpMockSpy;
+        beforeEach(() => {
+            bigIpMockSpy = sinon.spy(bigIpMock);
+        });
+
+        it('should have correct transaction command for creating new Route', () => {
+            state.currentConfig.Common.Route = {};
+            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
+            return networkHandler.process()
+                .then(() => {
+                    assert.strictEqual(bigIpMockSpy.create.callCount, 2);
+                    assert.strictEqual(bigIpMockSpy.delete.callCount, 0);
+                });
+        });
+
+        it('should have correct transaction commands for modifying existing Route', () => {
+            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
+            return networkHandler.process()
+                .then(() => {
+                    assert.strictEqual(bigIpMockSpy.create.callCount, 2);
+                    assert.strictEqual(bigIpMockSpy.delete.callCount, 2);
+                });
+        });
+
+        it('should correctly update network property with transactions', () => {
+            state.currentConfig.Common.Route = {
+                localRoute: {
+                    name: 'localRoute',
+                    gw: '10.11.12.13',
+                    network: '51.62.73.85/32',
+                    mtu: 1005,
+                    localOnly: true
+                }
+            };
+            declaration.Common.Route.localRoute.network = '51.62.73.86/32';
+            const networkHandler = new NetworkHandler(declaration, bigIpMock, null, state);
+            return networkHandler.process()
+                .then(() => {
+                    assert.strictEqual(bigIpMockSpy.create.callCount, 2);
+                    assert.strictEqual(bigIpMockSpy.delete.callCount, 1);
+                    assert.deepStrictEqual(deletedPaths, ['/tm/net/route/~LOCAL_ONLY~localRoute']);
+                    assert.strictEqual(dataSent[PATHS.Route][1].network, '51.62.73.86/32');
                 });
         });
     });
@@ -1025,6 +1006,122 @@ describe('networkHandler', () => {
                     assert.strictEqual(tunnels[1].usePmtu, 'disabled');
                     assert.strictEqual(tunnels[1].tos, 12);
                     assert.strictEqual(tunnels[1].autoLasthop, 'enabled');
+                });
+        });
+    });
+
+    describe('RoutingAsPath', () => {
+        it('should handle a fully specified Routing As Path', () => {
+            const declaration = {
+                Common: {
+                    RoutingAsPath: {
+                        RoutingAsPath1: {
+                            name: 'RoutingAsPath1',
+                            entries: [
+                                {
+                                    name: 10,
+                                    regex: '^$'
+                                }
+                            ]
+                        },
+                        RoutingAsPath2: {
+                            name: 'RoutingAsPath2',
+                            entries: [
+                                {
+                                    name: 15,
+                                    regex: 'funky$'
+                                },
+                                {
+                                    name: 20,
+                                    regex: '^$'
+                                }
+                            ]
+                        }
+                    }
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent[PATHS.RoutingAsPath];
+                    assert.deepStrictEqual(data[0], {
+                        name: 'RoutingAsPath1',
+                        partition: 'Common',
+                        entries: {
+                            10: {
+                                action: 'permit',
+                                regex: '^$'
+                            }
+                        }
+                    });
+                    assert.deepStrictEqual(data[1], {
+                        name: 'RoutingAsPath2',
+                        partition: 'Common',
+                        entries: {
+                            15: {
+                                action: 'permit',
+                                regex: 'funky$'
+                            },
+                            20: {
+                                action: 'permit',
+                                regex: '^$'
+                            }
+                        }
+                    });
+                });
+        });
+
+        it('should send out 1 enable value to the sys/db/...routing endpoint if a declaration includes RoutingAsPath', () => {
+            const declaration = {
+                Common: {
+                    RoutingAsPath: {
+                        RoutingAsPath1: {
+                            name: 'RoutingAsPath1',
+                            entries: [
+                                {
+                                    name: 10,
+                                    regex: '^$'
+                                }
+                            ]
+                        },
+                        RoutingAsPath2: {
+                            name: 'RoutingAsPath2',
+                            entries: [
+                                {
+                                    name: 15,
+                                    regex: 'funky$'
+                                },
+                                {
+                                    name: 20,
+                                    regex: '^$'
+                                }
+                            ]
+                        }
+                    }
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent['/tm/sys/db/tmrouted.tmos.routing'];
+                    assert.deepStrictEqual(data, [{ value: 'enable' }]);
+                });
+        });
+
+        it('should not send out an enable value to the sys/db/...routing endpoint if a declaration lacks RoutingAsPath', () => {
+            const declaration = {
+                Common: {
+                    RoutingAsPath: {}
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent['/tm/sys/db/tmrouted.tmos.routing'];
+                    assert.deepStrictEqual(data, undefined);
                 });
         });
     });